modify default maxRows type to Int, default client mode to cluster

+++ client/src/main/scala/moonbox/client/MoonboxClient.scala
@@ -46,8 +46,8 @@ trait MoonboxClient {
   def setReadTimeout(milliseconds: Int): Unit
   def getFetchSize: Int
   def setFetchSize(size: Int): Unit
-  def getMaxRows: Long
-  def setMaxRows(size: Long): Unit
+  def getMaxRows: Int
+  def setMaxRows(size: Int): Unit
   def getServers: Seq[InetSocketAddress]
   def getConf(key: String): Option[String]
   def getAllConf: java.util.Map[String, String]
@@ -80,7 +80,7 @@ trait MoonboxClient {
     * @return
     */
   def interactiveQuery(interactiveSql: Seq[String], fetchSize: Int, milliseconds: Int): MoonboxRowSet
-  def interactiveQuery(interactiveSql: Seq[String], fetchSize: Int, maxRows: Long, milliseconds: Int): MoonboxRowSet
+  def interactiveQuery(interactiveSql: Seq[String], fetchSize: Int, maxRows: Int, milliseconds: Int): MoonboxRowSet
   def cancelInteractiveQuery(): Boolean
   def cancelBatchQuery(jobId: String): Boolean
